Added collapse / expand group on mouse click and oher
Added collapse / expand group on mouse click.
fnGroupLabelFormat - Group name modification function. 
fnOnGroupCompleted - The function that is executed at the end of the formation of the start line of the group. Adds the number of lines in the group.
bExpandSingleGroup - variable is not used now.
sOpenGroup - the group that is displayed in expanded form
bHideGroupingColumn - variable is not used now.